pyproject.toml
[project]
name = "claude-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "httpx>=0.28.1",
    "requests>=2.32.0",
    "openai>=2.9.0",
    "aiohttp>=3.8.0",
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]

[[tool.uv.index]]
url = "https://mirror.nju.edu.cn/pypi/web/simple"
default = true
